x64: injectors: eliminate implicit narrowing conversions
(cherry picked from commit 39dbfc5f24c5e63161d532578753bba4d56c012b)

+++ src/cpu/x64/jit_uni_x8s8s32x_conv_kernel.cpp
@@ -63,14 +63,16 @@ jit_uni_x8s8s32x_fwd_kernel_vmm_t<isa, Vmm>::jit_uni_x8s8s32x_fwd_kernel_vmm_t(
         using namespace binary_injector;
         static constexpr bool preserve_gpr = true;
         static constexpr bool preserve_vmm = true;
-        static constexpr size_t helper_vmm_idx = 2;
-        const size_t block_tail
-                = (jcp.is_depthwise ? jcp.ch_block : jcp.oc_block)
-                % isa_simd_width_;
-        const size_t tail_size = block_tail
+        static constexpr int helper_vmm_idx = 2;
+        const int block_tail = static_cast<int>(
+            (jcp.is_depthwise ? jcp.ch_block : jcp.oc_block)
+            % isa_simd_width_);
+        const int tail_size = block_tail
                 ? block_tail
-                : (jcp.is_depthwise ? jcp.ngroups : jcp.oc_without_padding)
-                        % isa_simd_width_;
+            : static_cast<int>(
+                (jcp.is_depthwise ? jcp.ngroups
+                          : jcp.oc_without_padding)
+                % isa_simd_width_);
 
         const rhs_arg_static_params_t rhs_arg_static_params {helper_vmm_idx,
                 r13, r14, r15, preserve_gpr, preserve_vmm,
